Move collections out of the v2 directory because they are always the same implementation
