+++ src/models/v2/asyncapi.ts
@@ -1,15 +1,15 @@
 import { BaseModel } from '../base';
 import { Info } from './info';
-import { Channels } from './channels';
+import { Channels } from '../channels';
 import { Channel } from './channel';
 import { Components } from './components';
-import { Messages } from './messages';
-import { Operations } from './operations';
-import { Servers } from './servers';
+import { Messages } from '../messages';
+import { Operations } from '../operations';
+import { Servers } from '../servers';
 import { Server } from './server';
-import { SecuritySchemes } from './security-schemes';
+import { SecuritySchemes } from '../security-schemes';
 import { SecurityScheme } from './security-scheme';
-import { Schemas } from './schemas';
+import { Schemas } from '../schemas';
 
 import { extensions } from './mixins';
