+++ src/models/v2/bindings.ts
@@ -1,5 +1,5 @@
 import { Collection } from '../collection';
-import { Extensions } from './extensions';
+import { Extensions } from '../extensions';
 import { Extension } from './extension';
 
 import { createModel } from '../utils';
