--- src/models/v2/channel-parameters.ts
@@ -1,10 +1,0 @@
-import { Collection } from '../collection';
-
-import type { ChannelParametersInterface } from '../channel-parameters';
-import type { ChannelParameterInterface } from '../channel-parameter';
-
-export class ChannelParameters extends Collection<ChannelParameterInterface> implements ChannelParametersInterface {
-  override get(id: string): ChannelParameterInterface | undefined {
-    return this.collections.find(parameter => parameter.id() === id);
-  }
-}
