+++ src/models/v2/components.ts
@@ -13,17 +13,17 @@ import { SecurityScheme } from './security-scheme';
 import { Server } from './server';
 import { ServerVariable } from './server-variable';
 import { extensions } from './mixins';
-import { Servers } from './servers';
-import { Channels } from './channels';
-import { Messages } from './messages';
-import { Schemas } from './schemas';
-import { ChannelParameters } from './channel-parameters';
-import { ServerVariables } from './server-variables';
-import { OperationTraits } from './operation-traits';
-import { MessageTraits } from './message-traits';
-import { SecuritySchemes } from './security-schemes';
-import { CorrelationIds } from './correlation-ids';
-import { Operations } from './operations';
+import { Servers } from '../servers';
+import { Channels } from '../channels';
+import { Messages } from '../messages';
+import { Schemas } from '../schemas';
+import { ChannelParameters } from '../channel-parameters';
+import { ServerVariables } from '../server-variables';
+import { OperationTraits } from '../operation-traits';
+import { MessageTraits } from '../message-traits';
+import { SecuritySchemes } from '../security-schemes';
+import { CorrelationIds } from '../correlation-ids';
+import { Operations } from '../operations';
 import { Message } from './message';
 
 import { tilde } from '../../utils';
