+++ src/models/v2/message-trait.ts
@@ -1,6 +1,6 @@
 import { BaseModel } from '../base';
 import { CorrelationId } from './correlation-id';
-import { MessageExamples } from './message-examples';
+import { MessageExamples } from '../message-examples';
 import { MessageExample } from './message-example';
 import { Schema } from './schema';
