--- src/models/v2/server-variables.ts
@@ -1,10 +1,0 @@
-import { Collection } from '../collection';
-
-import type { ServerVariablesInterface } from '../server-variables';
-import type { ServerVariableInterface } from '../server-variable';
-
-export class ServerVariables extends Collection<ServerVariableInterface> implements ServerVariablesInterface {
-  override get(id: string): ServerVariableInterface | undefined {
-    return this.collections.find(variable => variable.id() === id);
-  }
-}
