+++ src/models/v2/server.ts
@@ -1,12 +1,12 @@
 import { BaseModel } from '../base';
-import { Channels } from './channels';
+import { Channels } from '../channels';
 import { Channel } from './channel';
-import { Messages } from './messages';
-import { Operations } from './operations';
+import { Messages } from '../messages';
+import { Operations } from '../operations';
 import { SecurityScheme } from './security-scheme';
-import { ServerVariables } from './server-variables';
+import { ServerVariables } from '../server-variables';
 import { ServerVariable } from './server-variable';
-import { SecurityRequirements } from './security-requirements';
+import { SecurityRequirements } from '../security-requirements';
 import { SecurityRequirement } from './security-requirement';
 
 import { bindings, hasDescription, description, extensions } from './mixins';
